refactor: 폴더트리 변화후 에러 수정
