+++ categories/selectors/selector_v1_category.py
@@ -1,6 +1,6 @@
 from django.db.models import QuerySet
 
-from categories.models import Category
+from categories.models.category import Category
 from utils.choices import CategoryKindChoices
