+++ medias/urls.py
@@ -1,5 +1,6 @@
 from django.urls import path
-from .views import PhotoDetail, GetUploadURL
+
+from medias.views import GetUploadURL, PhotoDetail
 
 urlpatterns = [
     path("photos/get-url", GetUploadURL.as_view()),
