+++ medias/views.py
@@ -5,7 +5,8 @@ from rest_framework.views import APIView
 from rest_framework.status import HTTP_200_OK
 from rest_framework.response import Response
 from rest_framework.exceptions import NotFound, PermissionDenied
-from .models import Photo
+
+from medias.models.photo import Photo
 
 
 class PhotoDetail(APIView):
